+++ Light.cpp
@@ -9,13 +9,13 @@ void DirectionalLight::lighting(const Material& material,
                                 glm::vec3& specular) const
 {
 
-    float diff = glm::max(glm::dot(normal, -this->m_direction), 0.0f);
+    double diff = glm::max(glm::dot(normal, -this->m_direction), 0.0f);
     glm::vec3 halfwayDir = eyeDir + this->m_direction;
     halfwayDir = glm::normalize(halfwayDir);
-    float spec = pow(glm::max(glm::dot(eyeDir, halfwayDir), 0.0f), material.m_shininess);
+    double spec = pow(glm::max(glm::dot(eyeDir, halfwayDir), 0.0f), material.m_shininess);
     ambient = m_ambient;
-    diffuse = m_diffuse * diff;
-    specular = m_specular * spec;
+    diffuse = m_diffuse * static_cast<float>(diff);
+    specular = m_specular * static_cast<float>(spec);
 }
 
 void PointLight::lighting(const Material& material,
@@ -33,18 +33,18 @@ void PointLight::lighting(const Material& material,
     glm::vec3 lightDir = (this->m_position - position);
     lightDir = glm::normalize(lightDir);
     
-    float diff = glm::max(glm::dot(normal, lightDir), 0.0f);
-    diffuse = this->m_diffuse * diff;
+    double diff = glm::max(glm::dot(normal, lightDir), 0.0f);
+    diffuse = this->m_diffuse * static_cast<float>(diff);
 
     // specular
     glm::vec3 halfwayDir = eyeDir + lightDir;
     halfwayDir = glm::normalize(halfwayDir);
-    float spec = pow(glm::max(glm::dot(eyeDir, halfwayDir), 0.0f), material.m_shininess);
-    specular = this->m_specular * spec;
+    double spec = pow(glm::max(glm::dot(eyeDir, halfwayDir), 0.0f), material.m_shininess);
+    specular = this->m_specular * static_cast<float>(spec);
 
     // attenuation
-    float distance = glm::length(this->m_position - position);
-    float attenuation = 1.0 / (m_attenuation.x +
+    double distance = glm::length(this->m_position - position);
+    double attenuation = 1.0 / (m_attenuation.x +
         m_attenuation.y * distance +
         m_attenuation.z * (distance * distance));
     ambient *= attenuation;
@@ -66,27 +66,27 @@ void SpotLight::lighting(const Material& material,
     // diffuse
     glm::vec3 lightDir = this->m_position - position;
     lightDir = glm::normalize(lightDir);
-    float diff = glm::max(glm::dot(normal, lightDir), 0.0f);
-    diffuse = this->m_diffuse * diff;
+    double diff = glm::max(glm::dot(normal, lightDir), 0.0f);
+    diffuse = this->m_diffuse * static_cast<float>(diff);
 
     // specular
     glm::vec3 halfwayDir = eyeDir + lightDir;
     halfwayDir = glm::normalize(halfwayDir);
-    float spec = pow(glm::max(glm::dot(eyeDir, halfwayDir), 0.0f), material.m_shininess);
-    specular = this->m_specular * spec;
+    double spec = pow(glm::max(glm::dot(eyeDir, halfwayDir), 0.0f), material.m_shininess);
+    specular = this->m_specular * static_cast<float>(spec);
 
     // spotlight (soft edges)
-    float theta = glm::dot(lightDir, -this->m_direction);
-    float epsilon = (this->m_cutoff - this->m_outcutoff);
-    float intensity = (theta - this->m_outcutoff) / epsilon;
+    double theta = glm::dot(lightDir, -this->m_direction);
+    double epsilon = (this->m_cutoff - this->m_outcutoff);
+    double intensity = (theta - this->m_outcutoff) / epsilon;
     if (intensity < 0.0f)intensity = 0.0f;
     if (intensity > 1.0f)intensity = 1.0f;
     diffuse *= intensity;
     specular *= intensity;
 
     // attenuation
-    float distance = glm::length(this->m_position - position);
-    float attenuation = 1.0 / (m_attenuation.x +
+    double distance = glm::length(this->m_position - position);
+    double attenuation = 1.0 / (m_attenuation.x +
         m_attenuation.y * distance +
         m_attenuation.z * (distance * distance));
     ambient *= attenuation;
